Update _to_png method to support custom size
Add parameters `use_custom_size` (to use custom dimensions for the export) and `use_map_size` (to use map dimensions for the export) to `_to_png` to add flexibility when exporting the map to an image. When none is set, default to the previous behavior by maximizing the webdriver window.